Refactor of all structure pieces
