+++ src/main/java/com/gildedgames/aether/world/structure/BronzeDungeonStructure.java
@@ -1,17 +1,15 @@
 package com.gildedgames.aether.world.structure;
 
 import com.gildedgames.aether.Aether;
-import com.gildedgames.aether.world.structurepiece.BronzeDungeonGraph;
+import com.gildedgames.aether.world.structurepiece.bronzedungeon.BronzeDungeonGraph;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Vec3i;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.LevelHeightAccessor;
 import net.minecraft.world.level.NoiseColumn;
 import net.minecraft.world.level.chunk.ChunkGenerator;
-import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.RandomState;
 import net.minecraft.world.level.levelgen.structure.BoundingBox;
 import net.minecraft.world.level.levelgen.structure.Structure;
