+++ src/main/java/com/gildedgames/aether/world/structurepiece/golddungeon/GoldDungeonPiece.java
@@ -1,0 +1,37 @@
+package com.gildedgames.aether.world.structurepiece.golddungeon;
+
+import com.gildedgames.aether.Aether;
+import com.gildedgames.aether.block.AetherBlocks;
+import com.gildedgames.aether.world.structurepiece.AetherTemplateStructurePiece;
+import com.google.common.collect.ImmutableList;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceType;
+import net.minecraft.world.level.levelgen.structure.templatesystem.*;
+
+import java.util.function.Function;
+
+/**
+ * Superclass for all gold dungeon structure pieces. This exists to simplify the code.
+ */
+public abstract class GoldDungeonPiece extends AetherTemplateStructurePiece {
+    public static final RuleProcessor LOCKED_HELLFIRE_STONE = new RuleProcessor(ImmutableList.of(
+            new ProcessorRule(new RandomBlockMatchTest(AetherBlocks.LOCKED_HELLFIRE_STONE.get(), 0.1F), AlwaysTrueTest.INSTANCE, AetherBlocks.LOCKED_LIGHT_HELLFIRE_STONE.get().defaultBlockState())
+    ));
+    public static final RuleProcessor MOSSY_HOLYSTONE = new RuleProcessor(ImmutableList.of(
+            new ProcessorRule(new RandomBlockMatchTest(AetherBlocks.HOLYSTONE.get(), 0.2F), AlwaysTrueTest.INSTANCE, AetherBlocks.MOSSY_HOLYSTONE.get().defaultBlockState())
+    ));
+
+    public GoldDungeonPiece(StructurePieceType type, StructureTemplateManager manager, String name, StructurePlaceSettings settings, BlockPos pos) {
+        super(type, manager, makeLocation(name), settings, pos);
+    }
+
+    public GoldDungeonPiece(StructurePieceType type, CompoundTag tag, StructureTemplateManager manager, Function<ResourceLocation, StructurePlaceSettings> settingsFactory) {
+        super(type, tag, manager, settingsFactory);
+    }
+
+    protected static ResourceLocation makeLocation(String name) {
+        return new ResourceLocation(Aether.MODID, "gold_dungeon/" + name);
+    }
+}
