+++ src/main/java/com/gildedgames/aether/world/structurepiece/silverdungeon/SilverBossRoom.java
@@ -1,0 +1,83 @@
+package com.gildedgames.aether.world.structurepiece.silverdungeon;
+
+
+import com.gildedgames.aether.Aether;
+import com.gildedgames.aether.api.DungeonTracker;
+import com.gildedgames.aether.blockentity.TreasureChestBlockEntity;
+import com.gildedgames.aether.data.resources.registries.AetherStructures;
+import com.gildedgames.aether.entity.AetherEntityTypes;
+import com.gildedgames.aether.entity.monster.dungeon.boss.ValkyrieQueen;
+import com.gildedgames.aether.loot.AetherLoot;
+import com.gildedgames.aether.world.structurepiece.AetherStructurePieceTypes;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.level.ServerLevelAccessor;
+import net.minecraft.world.level.StructureManager;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.Rotation;
+import net.minecraft.world.level.block.entity.RandomizableContainerBlockEntity;
+import net.minecraft.world.level.levelgen.structure.BoundingBox;
+import net.minecraft.world.level.levelgen.structure.Structure;
+import net.minecraft.world.level.levelgen.structure.StructureStart;
+import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceSerializationContext;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructurePlaceSettings;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplateManager;
+import net.minecraft.world.phys.AABB;
+
+import java.util.ArrayList;
+
+public class SilverBossRoom extends SilverDungeonPiece {
+
+    public SilverBossRoom(StructureTemplateManager manager, String name, BlockPos pos, Rotation rotation) {
+        super(AetherStructurePieceTypes.SILVER_BOSS_ROOM.get(), manager, name, makeSettings().setRotation(rotation), pos);
+        this.setOrientation(rotation.rotate(Direction.SOUTH));
+    }
+
+    public SilverBossRoom(StructurePieceSerializationContext context, CompoundTag tag) {
+        super(AetherStructurePieceTypes.SILVER_BOSS_ROOM.get(), tag, context.structureTemplateManager(), resourceLocation -> makeSettings());
+    }
+
+    private static StructurePlaceSettings makeSettings() {
+        return new StructurePlaceSettings().addProcessor(LOCKED_ANGELIC_STONE);
+    }
+
+    @Override
+    protected void handleDataMarker(String name, BlockPos pos, ServerLevelAccessor level, RandomSource random, BoundingBox chunkBB) {
+        if (name.equals("Valkyrie Queen")) {
+            ValkyrieQueen queen = new ValkyrieQueen(AetherEntityTypes.VALKYRIE_QUEEN.get(), level.getLevel());
+            queen.setPersistenceRequired();
+            queen.setPos(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5);
+            queen.setDungeon(new DungeonTracker<>(queen,
+                    queen.position(),
+                    new AABB(this.boundingBox.minX(), this.boundingBox.minY(), this.boundingBox.minZ(), this.boundingBox.maxX() + 1, this.boundingBox.maxY() + 1, this.boundingBox.maxZ() + 1),
+                    new ArrayList<>()));
+            // Set the bounds for the whole dungeon
+            StructureManager manager = level.getLevel().structureManager();
+            manager.registryAccess().registry(Registries.STRUCTURE).ifPresent(registry -> {
+                        Structure temple = registry.get(AetherStructures.SILVER_DUNGEON);
+                        if (temple != null) {
+                            StructureStart start = manager.getStructureAt(pos, temple);
+                            if (start != StructureStart.INVALID_START) {
+                                BoundingBox box = start.getBoundingBox();
+                                AABB dungeonBounds = new AABB(box.minX(), box.minY(), box.minZ(), box.maxX() + 1, box.maxY() + 1, box.maxZ() + 1);
+                                queen.setDungeonBounds(dungeonBounds);
+                            }
+                        }
+                    }
+            );
+            queen.finalizeSpawn(level, level.getCurrentDifficultyAt(pos), MobSpawnType.STRUCTURE, null, null);
+            level.getLevel().addFreshEntity(queen);
+            level.setBlock(pos, Blocks.AIR.defaultBlockState(), 2);
+        } else if (name.equals("Treasure Chest")) {
+            BlockPos chest = pos.below();
+            RandomizableContainerBlockEntity.setLootTable(level, random, chest, AetherLoot.SILVER_DUNGEON_REWARD);
+            TreasureChestBlockEntity.setDungeonType(level, chest, new ResourceLocation(Aether.MODID, "silver"));
+            level.setBlock(pos, Blocks.AIR.defaultBlockState(), 2);
+        }
+    }
+}
